requirements.txt
fastapi
uvicorn
chromadb
google-genai
sentence-transformers
pydantic
tqdm
